First code drop for git-decent
This commit contains:
- config code which retrieves the configuration from git
- schedule which parses the config and returns when making commits is ok
- basic git repository management and test infra

Transitioned from a solely go-git approach to a hybrid with git commands
due to challenges in commit signing. Given that git-decent is an
extension to git, it's reasonable to expect git to be installed and
available at $PATH. It's also justifiable to assume correct gpg or ssh
signing configuration if amending a signed commit.