Add the concept of TimeFrames in schedule
Every time range defiend in the conflict is a "Decent" timeframe where
commits are acceptable.

So now the 1440 minutes of a day are a pointer to a TimeFrame or nil if
they are not decent minutes.

Also TimeFrames are linked to the next TimeFrame from the same day, this
way we can easily go trhough available minutes even across TimeFrames.

It might not be the most efficient data structure but this singly linked
list should be fast enough.