Add WithCommitsWithDates to fixtures builder
In order to finally implement the app logic we need an easy way of
showing how the dates will be amended, a sorts of before and after.

This method accepts a list of time.Time that will be used to create
commits with all random data but the date.

This way we can generate logs to amend.