Implement more of the Amend logic
This is just a first implementation to discover exactly what we want and
what we need. The function has different reponsabilities, the algorithm
is implemented in a hackish way, etc etc.

We will rewrite this once we've all the tests in place and properly abastract
whatever needs to be abstracetd.