Add schedule.ClosestDecentMinute
Something we've learned in the first Amend PoC implementation is that we want to deal
with minutes instead of days or even frames. We want to know what the next decent
mintue is, that's it.

So this commit adds such as method.

If we end up only using minutes, we will have to refactor schedule to better fit
that.