Reimplement the Amending logic
With what we learned implementing the first PoC, rewritting it has been a breeze.

The data in the table fixtures have been modified since now the amending logic is more accurate
than before, for example

In the Amend Overflow case we were allowing the amended date to be a little bit after the
decent frame, now that is no longuer allwoed and the date is proerly moved to the next
decent frame.