Add NewScheduleFromPlainText
The first time git-decent is executed we allow the user to
configure the configuration which is done in plain text.

Before we copy this new configuration to .git/config we want
to validate it and to do so it is just easier if we can skip
adding it to .git/config just to delete it afterwards.

So this commit adds NewScheduleFromPlainText which accepts the
entire git config section as plaintext, converts it to a
RawSchedule and calls NewScheduleFromRaw.

We haven't minded the performance at all since this should just
be called once.