Add a sort of first time wizard
It will check if a schedule exists and if it does
not it will open the git editor with a template that
shows how the configuration syntax works.

Then it pretty prints it.

Most probably we want to add this before any git decent
commadn is executed but for now it lives in the root command.