Also amend the commit date
Each commit has two dates, one is when it was authored
and the other is when the commit was created or edited.

For example git commit --amend will keep the Author date
but modify the commit date.

This is not really required for git-decent since our
objective is to be inconspicuous not to be completly
sealthy so amending the Author date is enough but seeing
different dates can be confusing and users might assume that
the tool did not work correctly, thus we are also changing it.