Add hook command
This is just a PoC copy pasting code from the rootCommand.

We've learned a few things like some code that we should be
able to share between hook, amendlast and amendUnpushed.

Also the usage of PreRun is not really working out for us
since the error reporting does not look good, further
investigation is needed

Eventually we will have a command to just amend the head, that command
will be public and docummented while hook is kept private.

The only reason why hookCmd exists is to avoid a possible amending
infinite loop

git commit --> post-commit hook --> git amend --> post-commit hook...

This loop shouldn't happen since git-decent amending should change the
date in a way that next time it is executed it shouldn't change anything
but it is better to be safe than sorry.