Prevent hook executing if Git state is not clean
The post commit hook is going to be executed everytime a
commit is created which happens everytime a commit is edited
for exampel by changing its father.

If we change the commit in the middle of a merge or rebase
we can cause unintended consequences.

Since we are already abusing post-commit hook for git decent
let's try to be as a good of a citizen as we can and just do
nothing if the state is not clean.