Prevent git-decent from havint itself as children process
Since we are using git hooks it can happen that we end up executing
git-decent from git-decent itself, for example

git decent hooks amend

This will
[git commit --amend] --> [post-commit-hook] --> [git-decent hooks amend]...

We already have some fallbacks in place but what we did not had was a happy
path to handle this.

Well, the happy path has now been implemented using a simple Env which git forwards
when executing hooks.

We've kept all the other checks like the lock-file just in case git will ever
stop forwarding the envs.