Print errros only once in PreRun and RunE
We are starting to refactor the app so only the most
external layers of the app deal with print wihle the rest
is kept as ui-free as possible.

This is the first step, we've disabled cobra built in
error printing and now all the printing will be done by
the Execute function in the cmd package.