Refactor debug ui
We've changed the ui package so it no longuer export
the Ui struct but instead can be used directly. This
has been done so we do not need to pass the struct
around and instead we can just use the ui package which
we believe, is a good case for a defacto singleton.

This commit also contains some other small miscelnaea
like adding the concept of DebugBlock that will print
a bunch of debug lines with a tittle as a block. It is
handy to see things like executions of the git command.