Add the concept of threshold
When it is not 0, the amender will keep the interval
between two commits as long as they do not exceed the
amount of minutes defined in threshold.

This allows for this bit of history to be kept intact
as long as commits do not cross the bounderies between
decent frames.

So for example, let's say 3 commits are done in intervals
of 10 minutes and all three need to be amended.

If the threshold is greater than 10 minutes they the interval
of 10 minutes will be kept as long as they all fit in one
decent frame.